Devops/Kubernetes/FLASK-REDIS-MONGODB/requirements.txt
flask 
redis 
Flask-PyMongo